Add prof-fix and remove prona blast run
Addresses issues:
https://git.rostlab.org/predictprotein/predictprotein/-/issues/6
https://git.rostlab.org/predictprotein/predictprotein/-/issues/4

The output of prof was outputting an 'L' in its results if
something was unknown. Now, it will output a question mark '?'.

ProNA (predbind) was running blast against big80; however, it is
already run in the pipeline, making it unnecessary.

`prof-fix` git submodule has been added - see the `.gitmodules` file
for details.